+++ api/package.json
@@ -20,7 +20,6 @@
     "express": "^4.18.2",
     "express-validator": "^7.0.1",
     "jsonwebtoken": "^9.0.0",
-    "mongoose": "^7.0.1",
-    "node-cron": "^3.0.2"
+    "mongoose": "^7.0.1"
   }
 }
