+++ api/src/helpers/Index.js
@@ -2,5 +2,6 @@ import { checkMinAmountProduct } from './CheckMinAmountProduct.js'
 import { generateJWT } from './GenerateJwt.js'
 import { generateNewFolio } from './FoliosGenerator.js'
 import { makeObjectNotification } from './MakeObjectNotification.js'
+import { checkVencimientoCxC, checkVencimientoCxP } from './Notificacion.js'
 
-export { checkMinAmountProduct, generateJWT, generateNewFolio, makeObjectNotification }
+export { checkMinAmountProduct, generateJWT, generateNewFolio, makeObjectNotification, checkVencimientoCxC, checkVencimientoCxP }
