+++ api/src/models/Cuentas_por_Pagar.js
@@ -67,6 +67,10 @@ const CUENTASxPAGAR = new Schema(
       type: String,
       default: 'MXN'
     },
+    NOTIFICADA: {
+      type: Boolean,
+      default: false
+    }
   },
   { timestamps: true }
 )
